src/main/java/com/thoughtworks/tdd/parklinglot/shell/controller/GotoParkingController.java
package com.thoughtworks.tdd.parklinglot.shell.controller;

import com.thoughtworks.tdd.parklinglot.core.ParkingBoy;
import com.thoughtworks.tdd.parklinglot.shell.io.Request;
import com.thoughtworks.tdd.parklinglot.shell.io.Response;

public class GotoParkingController implements BaseController {
    private final Request request;
    private final Response response;
    private ParkingBoy parkingBoy;

    public GotoParkingController(Request request, Response response, ParkingBoy parkingBoy) {

        this.request = request;
        this.response = response;
        this.parkingBoy = parkingBoy;
    }

    @Override
    public String process() {

        String forwardPath = "";
        if (parkingBoy.isFull()) {
            response.send("车库已满，无法停车！");
            forwardPath = "forward:main";
        } else {
            response.send("请输入车牌号:");
        }
        return forwardPath;
    }
}
